youtube-translator-server-main/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.13.3
blinker==1.9.0
boto3==1.36.18
botocore==1.36.18
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
deep-translator==1.11.4
distro==1.9.0
ffmpeg==1.4
Flask==3.1.0
Flask-Cors==5.0.0
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
MarkupSafe==3.0.2
openai==1.61.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.11.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
yt-dlp==2025.1.26
python-dotenv==1.0.1